+++ helpers/index.js
@@ -1,3 +1,4 @@
 const {HttpError} = require("./HttpError");
+const {ctrlWrapper} = require("./ctrlWrapper");
 
-module.exports = { HttpError };
+module.exports = { HttpError, ctrlWrapper };
