+++ routes/api/contacts.js
@@ -1,107 +1,24 @@
 const express = require("express");
-const { HttpError } = require("../../helpers/index");
-const Joi = require("joi");
+
 
 const {
-  listContacts,
-  getContactById,
-  addContact,
-  removeContact,
-  updateContact,
-} = require("../../models/contacts.js");
+  getAll,
+  getById,
+  add,
+  updateById,
+  deleteById,
+} = require("../../controllers/contacts");
 
 const router = express.Router();
 
-const addSchema = Joi.object({
-  id: Joi.string().required(),
-  name: Joi.string().required(),
-  phone: Joi.string().required(),
-  email: Joi.string().required(),
-});
+router.get("/", getAll);
 
-const updateSchema = Joi.object({
-  id: Joi.string(),
-  name: Joi.string(),
-  phone: Joi.string(),
-  email: Joi.string(),
-});
+router.get("/:contactId", getById);
 
-router.get("/", async (req, res, next) => {
-  try {
-    const data = await listContacts();
-    res.json(data);
-  } catch (error) {
-    next(error);
-  }
-});
+router.post("/", add);
 
-router.get("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const data = await getContactById(contactId);
+router.put("/:contactId", updateById);
 
-    if (!data) {
-      throw HttpError(404, "Not Found");
-    }
-
-    res.json(data);
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body);
-
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-
-    const data = await addContact(req.body);
-    res.status(201).json(data);
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.put("/:contactId", async (req, res, next) => {
-  try {
-    const { error } = updateSchema.validate(req.body);
-
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-
-    const {
-      params: { contactId },
-      body,
-    } = req;
-
-    const data = await updateContact(contactId, body);
-
-    if (!data) {
-      throw HttpError(404, "Not Found");
-    }
-    res.json(data);
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.delete("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const data = await removeContact(contactId);
-
-    if (!data) {
-      throw HttpError(404, "Not Found");
-    }
-    
-    res.json(data);
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete("/:contactId", deleteById);
 
 module.exports = router;
