+++ server.js
@@ -7,13 +7,12 @@ const PORT = 3001;
 const { PASSWORD, DB_USERNAME, CLASTERNAME, DB_NAME } = process.env;
 
 const DB_HOST = `mongodb+srv://${DB_USERNAME}:${PASSWORD}@${CLASTERNAME}.ohx0zsm.mongodb.net/${DB_NAME}?retryWrites=true&w=majority`;
-console.log(process.env);
 mongoose.set("strictQuery", true);
 
 mongoose
   .connect(DB_HOST)
   .then(() => {
-    console.log("connect seccesfull");
+    console.log("connect succesfull");
     app.listen(PORT, () => {
       console.log(`Server running. Use our API on port: ${PORT}`);
     });
